added utils file for cleaner traduction  script
